// src/types/CheckoutInterface.ts
export interface CheckoutCompanyDetailsInterface {
  ReturnChallanNumber: string | null;
  SaleChallanNumber: string | null;
  FocChallanNumber: string | null;
  CompanyFromPanNumber: string | null;
  CompanyToPanNumber: string | null;
  CompanyFrom: string;
  CompanyTo: string;
  CompanyFromAddress: string;
  CompanyToAddress: string;
  CompanyFromPhoneNumber: string;
  CompanyToPhoneNumber: string;
  CompanyFromId: string;
  CompanyToGUId: string;
}
// ...
export interface CheckoutBarcodeListInterface {
  TotalQuantity: number;
  ChallanNumber: string;
  BarcodeDetails: {
    Product: string;
    PackSize: string;
    Barcode: string;
    Quantity: number;
  }[];
}